Actually, let's let the garbage collector manage file descriptors
In just a moment, we're going to be spawning threads to handle
evaluation of subtrees.  So, we can't rely on scope or scope-like things
to accurately reflect the lifetime of the fd.  I'm a little concerned
about the gc not running often enough and running out of fds, but its
easy enough to work around that.

- We can just gc before every command (yuck)
- We can gc when an open/pipe command fails (eh?)
- We can switch to a ref counted model (eeeh...)