fds are ACTUALLY managed correctly, now
I swear!

Squashing was only good for creating a map of their fd -> our fd.  It
wasn't useful for handing off to fork-exec.  However, we only had to
squash because we were using an alist instead of a hash table.  We only
used an alist because it made fd scopes cheap.  With fset, we can drop
squashing and keep our cheap fd scopes.