Almost support building on ECL
ECL has a bug that prevents this from building.  When you build a
monolithic library, ECL doesn't properly link in the dependencies.  The
resulting .a contains .o files (for the parts of SHCL) and .a files (for
the libraries that SHCL depends on).  This is malformed and results in
errors at link time.

I'm not TOO concerned about supporting ECL, since it has a few fatal
flaws.
1. SHCL takes forever to boot (I worked around the issue above).  I
think that the dependencies that SHCL pulls in results in very long load
times.  This is pretty unavoidable.
2. ECL listens to SIGCHLD and automatically reaps all children.  This
makes it incompatible with SHCL's need to wait for children.  It looks
like this might be configurable (as an ECL boot option), but it is
unclear whether that option actually changes anything.

Mostly, building and (maybe, one day) testing under ECL is a good way to
prove that SHCL is portable.  So, don't expect to run SHCL on ECL any
time soon.