Make the fd-table lock recursive
Also, did you know we weren't holding the lock in retain-fd-bindings!?

Consider the following sequence of events
1. Bind *error-output* to a symbolic fd stream
2. Call open-retained (or any other lock-holding function that logs)
3. Unlike they do currently, assume the log functions write to *error-output*

When the symbolic fd stream tries to look up its concrete fd, it will
try to acquire the lock held by open-retained.

These functions should generally be able to handle sharing the data
structures with functions they call.  The one questionable one is
with-living-fds.  If you enter a new fd into the table in the body of
that function, your list won't contain that fd.  Uh oh.