Stop changing the process directory when forking
This requires having our own version of posix_spawn.  So,
libshcl-support is now required all the time.  Unfortunately, I was only
able to build things sanely with tup if it was able to run the build job
in a chroot (since shcl, when it gets revived, needs to find
libshcl-support in the same place it was when it was dumped).  Since
that has some unusual requirements (suid root or special kernel
features), I decided its time to stop using tup.  :(