Try to prevent re-compiling updaters for data classes
I'm not sure if there's a better time to evaluate the
define-cloning-accessor form.  I've tried to find one, but I can't.  The
problem is that the direct slot definition is created twice (if you
compile and then load).  Sucks.  Let's try to prevent the style warnings
by only defining the accessor if we haven't already (or if we defined it
incompatibly previously).