Refactor processing of glob patterns
This was necessary because the old make-wild-path-from-fragments was
very complex.  Adding support for [] glob patterns was nigh impossible.
Under the new model, it should be MUCH easier.  As proof, I threw
together a super hacky (not right at all) implementation.  Anyway, Now
there are 4 stages.

1. We "split" the fragment sequence into sequences of fragments that don't
contain #\/.  These are called "segments".  These are basically just
sequence of fragments describing a single part of the path.  This is
done by split-fragments-into-segments

2. Normally, empty segments represent consecutive slashes.  Empty
segments at the start of the segment sequence represent leading slashes.
In step 2, we deal with the special case of leading slashes and strip
off segments that resulted from leading slashes.  This is done by
handle-leading-slashes.

3. For each remaining segments left over from step 2, construct a
wild-path component (string or regex scanner) that describes the
matching behavior for that path component.  This is done by
make-wild-path-component-from-fragments.

4. Assemble the leading slashes from step 2 and the wild path components
from step 3 into a wild path.  This is done by
make-wild-path-from-fragments.