Switch to a simpler, more correct version of as->
I admit that I thought it was cool that as-> could be used to build up
arbitrary expressions (e.g. (as-> 'lambda x (x (arg) arg))), but
that's not really something I want to encourage.  Most of the time,
you want to thread values.  In the rare case where you DO want to
thread forms, you can find another solution.