Expose a way to reach into a readtable and get an inner readtable
This isn't very useful (yet!).  Readtables don't enforce that you only
deal in characters.  So, it is possible to nominate a "character
sequence" that includes, for example, a symbol.  We're going to use
this to support having extensible syntax inside double-quote contexts!
When reading a double-quote, we'll pull the double-quote readtable out
of the overall readtable and start reading extensibly.  Fancy!