Remove :updater from define-data, greatly simplify cloning accessors
The old configuration made it awkward to document the reader functions
and very awkward if a subclass wanted to override an updater.  It was
also a bit annoying that there wasn't a good way to update two slots
at once.  Now you can clone and use unsafe setters (or reinitialize
the instance).