Rejigger how completion suggestions are provided
This makes it easier for custom completers to be provided... I think.

Note that when expanding types for tab complete, we don't descend into
reserved words. I mean, strictly speaking, all the reserved words are
valid completion suggestions if we're looking for a-word.  However,
they're probably not the completions you're looking for when you want
an a-word.