Factor expansion-for-words a bit
Now its easier to test individual parts of the expansion pipeline.  It
will also help with tab complete, I think.

Tab complete has some requirements for expansion logic.  If the first
word is "foo\*", then we'll get that as a compound-word with two
parts: a simple word ("foo"), and a quoted word ("*").  If the user
hits tab after the #\*, we need to know what the "expanded" string
representation is (in this case, "foo*"), but we also need to know
that the #\* is quoted.  If we're going to look for a filename
completion candidate, we need to know that the user meant the literal
asterisk character instead of a globbing asterisk character.

So, tab complete is going to need to hook into expansion in
interesting ways.  I'm not quite sure what that will look like, but I
think that decomposing expansion into simpler parts will help.