Load libshcl-support lazily
At build time libshcl-support won't be in its final resting place.  If
we try to load it at build time, then we'll try to re-load it from the
same path at revival time.  Uh oh.  If we just wait to load it until
later, we may actually be able to find it in the right place.