Introduce a new parser
Its (almost) a complete re-write with new semantics.  To make it
easier to develop, test, and understand we're introducing it as a new
file initially.  Next we'll make the shell grammar and tab complete
interact with this parser.  Finally, we'll remove the old parser
entirely.

This parser is based on walkables instead of iterators and special
variables instead of explicitly passing the sequence around.  It also
has built-in tracking of parse errors (unlike the previous parser).

I tried very, very hard to keep parser state lexical instead of
special (so parser functions wouldn't have side effects), but
ultimately decided that it wasn't worth the complexity.  I was more
than happy to make the parser complicated, but I couldn't hide the
complexity from users of the parser in a satisfying way.  This style
of parser feels a lot more natural to use.

...

Also, the implementation is much simpler!  So... there's that.