Move everyone over to the new parser
Also, while we're in shcl/shell/complete, let's update it a bit to use
iterators less.  We could theoretically keep using iterators as the
currency type in shell/complete, but it gets awkward when we're
juggling both iterators and sequences.  Its better to move most of the
package over to sequences.