Freeze osx version of SBCL for CI builds
It looks like SBCL-1.5.3 is emitting potentially illegitimate
compilation errors.  I need to investigate a bit, but for the time
being let's stick to what works.

Unfortunately, 1.5.2 still seems to have the -no-pie issue that caused
us to freeze the linux version a while back.  I don't see that failure
in the 1.5.3 builds.  Once we get past this new issue we might be able
to build with latest SBCL again!