Update the ubuntu build for CI
The -no-pie build failure is caused by GCC being super old.  Let's
update it!

Interestingly, travis limits the set of available operating systems
based on what the language supports.  Since common-lisp is
unrecognized, it was limited to the default OS.  In order to use a
more recent Ubuntu variant, I needed to drop the language declaration.